refactor worker runner class
